requirements.txt
ipython==8.12.0
matplotlib==3.4.3
networkx==2.8.4
numpy==1.23.3
openmesh==1.2.1
pandas==1.4.3
PySide6>=6.5.0
PyWavefront==1.3.3
scikit_learn==1.2.2
scipy==1.9.1
tinyobjloader==2.0.0rc6
tqdm==4.62.3
trimesh==3.12.7
networkx>=1.2.0
tqdm>=4.62.3
trimesh>=3.12.7
torch==1.13.1+cu117
torch_geometric==2.3.0
torch_scatter==2.1.1+pt113cu117
torchmeta==1.8.0
psbody_mesh==0.3
